Move libs/apps code into libs/apps/runlocal subdirectory
The code in libs/apps was specific to the 'apps run-local' command.
Refactored it into libs/apps/runlocal to make libs/apps a collection
of packages rather than a single package.

- Moved all files from libs/apps/ to libs/apps/runlocal/
- Updated package name from 'apps' to 'runlocal'
- Updated import in cmd/workspace/apps/run_local.go
