+++ libs/apps/runlocal/env.go
@@ -1,4 +1,4 @@
-package apps
+package runlocal
 
 import (
 	"strconv"
