+++ libs/apps/runlocal/node.go
@@ -1,4 +1,4 @@
-package apps
+package runlocal
 
 import (
 	"context"
