+++ libs/apps/runlocal/python.go
@@ -1,4 +1,4 @@
-package apps
+package runlocal
 
 import (
 	"context"
